+++ visaservice/vst/pkg/packets/generator.go
@@ -1,4 +1,4 @@
-package conform
+package packets
 
 import (
 	"fmt"
