+++ visaservice/vst/pkg/testfw/api.go
@@ -1,0 +1,6 @@
+package testfw
+
+type Tester interface {
+	Name() string
+	Run(state *TestState, ctest *TestRun) error
+}
