+++ visaservice/vst/pkg/zcrypt/hmac.go
@@ -1,24 +1,14 @@
-package conform
+package zcrypt
 
 import (
 	"bytes"
 	"crypto/sha256"
-	"crypto/x509"
 	"encoding/binary"
-	"encoding/pem"
 )
-
-func certToPEM(cert *x509.Certificate) []byte {
-	blk := pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: cert.Raw,
-	}
-	return pem.EncodeToMemory(&blk)
-}
 
 // Milestone2 no-crypto hmac value.
 // See libnode/src/m2.rs
-func newM2HMAC(challengeData []byte, sessionID int32, timestamp int64) []byte {
+func GenM2HMAC(challengeData []byte, sessionID int32, timestamp int64) []byte {
 	var buf bytes.Buffer
 	buf.Write(challengeData)
 	binary.Write(&buf, binary.BigEndian, uint64(timestamp))
